feat: Add processor class BuildCommandsFromIds to AppStateProcessor
This code was present in CLI.cpp. It is now separated and implemented in AppStateProcessor so that also the GUI can make use of it

style: replace unordered_map with map so that workflow's steps can be iterated in sorted order (by their IDs, here keys of the maps)

feat: class Command implements getName(). This method returns the name of the workflow step, no matter the type of the processor (either raw data or sequence segment)